+++ .github/workflows/ci.yml
@@ -6,16 +6,20 @@ on:
 
 jobs:
   validate:
-    runs-on: ubuntu-latest
+    runs-on: [self-hosted, linux, x64]
 
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
 
+      - name: Clean stale artifacts
+        run: rm -rf node_modules ui/node_modules dist
+
       - name: Setup Bun
         uses: oven-sh/setup-bun@v2
         with:
           bun-version: '1.3.9'
+          no-cache: true
 
       - name: Setup Node.js
         uses: actions/setup-node@v4
