--- src/commands/api-command.ts
@@ -1,1 +1,0 @@
-export * from './api-command/index';
